kernel: give each process a unique applet state
