+++ src/core/horizon/kernel/const.hpp
@@ -484,6 +484,53 @@ enum class ArbitrationType {
     WaitIfEqual = 2,
 };
 
+enum class AppletMessage {
+    None = 0,
+    ChangeIntoForeground = 1,
+    ChangeIntoBackground = 2,
+    Exit = 4,
+    ApplicationExited = 6,
+    FocusStateChanged = 15,
+    Resume = 16,
+    DetectShortPressingHomeButton = 20,
+    DetectLongPressingHomeButton = 21,
+    DetectShortPressingPowerButton = 22,
+    DetectMiddlePressingPowerButton = 23,
+    DetectLongPressingPowerButton = 24,
+    RequestToPrepareSleep = 25,
+    FinishedSleepSequence = 26,
+    SleepRequiredByHighTemperature = 27,
+    SleepRequiredByLowBattery = 28,
+    AutoPowerDown = 29,
+    OperationModeChanged = 30,
+    PerformanceModeChanged = 31,
+    DetectReceivingCecSystemStandby = 32,
+    SdCardRemoved = 33,
+    LaunchApplicationRequested = 50,
+    RequestToDisplay = 51,
+    ShowApplicationLogo = 55,
+    HideApplicationLogo = 56,
+    ForceHideApplicationLogo = 57,
+    FloatingApplicationDetected = 60,
+    DetectShortPressingCaptureButton = 90,
+    AlbumScreenShotTaken = 92,
+    AlbumRecordingSaved = 93,
+    DetectShortPressingCaptureButtonForApplet = 110,
+    DetectLongPressingCaptureButtonForApplet = 111,
+};
+
+enum class AppletFocusState {
+    InFocus = 1,    ///< Applet is focused.
+    OutOfFocus = 2, ///< Out of focus - LibraryApplet open.
+    Background = 3  ///< Out of focus - HOME menu open / console is sleeping.
+};
+
+enum class LaunchParameterKind : u32 {
+    UserChannel = 1,
+    PreselectedUser = 2,
+    Unknown0 = 3,
+};
+
 } // namespace hydra::horizon::kernel
 
 ENABLE_ENUM_FORMATTING(
@@ -597,3 +644,36 @@ ENABLE_ENUM_FORMATTING(hydra::horizon::kernel::ArbitrationType, WaitIfLessThan,
                        "wait if less than", DecrementAndWaitIfLessThan,
                        "decrement and wait if less than", WaitIfEqual,
                        "wait if equal")
+
+ENABLE_ENUM_FORMATTING(
+    hydra::horizon::kernel::AppletMessage, None, "none", ChangeIntoForeground,
+    "change into foreground", ChangeIntoBackground, "change into background",
+    Exit, "exit", ApplicationExited, "application exited", FocusStateChanged,
+    "focus state changed", Resume, "resume", DetectShortPressingHomeButton,
+    "detect short pressing home button", DetectLongPressingHomeButton,
+    "detect long pressing home button", DetectShortPressingPowerButton,
+    "detect short pressing power button", DetectMiddlePressingPowerButton,
+    "detect middle pressing power button", DetectLongPressingPowerButton,
+    "detect long pressing power button", RequestToPrepareSleep,
+    "request to prepare sleep", FinishedSleepSequence,
+    "finished sleep sequence", SleepRequiredByHighTemperature,
+    "sleep required by high temperature", SleepRequiredByLowBattery,
+    "sleep required by low battery", AutoPowerDown, "auto power down",
+    OperationModeChanged, "operation mode changed", PerformanceModeChanged,
+    "performance mode changed", DetectReceivingCecSystemStandby,
+    "detect receiving cec system standby", SdCardRemoved, "sd card removed",
+    LaunchApplicationRequested, "launch application requested",
+    RequestToDisplay, "request to display", ShowApplicationLogo,
+    "show application logo", HideApplicationLogo, "hide application logo",
+    ForceHideApplicationLogo, "force hide application logo",
+    FloatingApplicationDetected, "floating application detected",
+    DetectShortPressingCaptureButton, "detect short pressing capture button",
+    AlbumScreenShotTaken, "album screenshot taken", AlbumRecordingSaved,
+    "album recording saved", DetectShortPressingCaptureButtonForApplet,
+    "detect short pressing capture button for applet",
+    DetectLongPressingCaptureButtonForApplet,
+    "detect long pressing capture button for applet")
+
+ENABLE_ENUM_FORMATTING(hydra::horizon::kernel::LaunchParameterKind, UserChannel,
+                       "user channel", PreselectedUser, "preselected user",
+                       Unknown0, "unknown0")
