+++ src/core/horizon/services/am/application_functions.cpp
@@ -2,7 +2,6 @@
 
 #include "core/horizon/kernel/process.hpp"
 #include "core/horizon/services/am/storage.hpp"
-#include "core/horizon/state_manager.hpp"
 
 namespace hydra::horizon::services::am {
 
@@ -16,12 +15,14 @@ DEFINE_SERVICE_COMMAND_TABLE(IApplicationFunctions, 1, PopLaunchParameter, 20,
                              EnableApplicationCrashReport, 130,
                              GetGpuErrorDetectedSystemEvent)
 
-result_t IApplicationFunctions::PopLaunchParameter(add_service_fn_t add_service,
-                                                   LaunchParameterKind kind) {
+result_t
+IApplicationFunctions::PopLaunchParameter(kernel::Process* process,
+                                          add_service_fn_t add_service,
+                                          kernel::LaunchParameterKind kind) {
     LOG_DEBUG(Services, "Kind: {}", kind);
 
     add_service(
-        new IStorage(StateManager::GetInstance().PopLaunchParameter(kind)));
+        new IStorage(process->GetAppletState().PopLaunchParameter(kind)));
     return RESULT_SUCCESS;
 }
