+++ src/core/horizon/services/am/application_functions.hpp
@@ -23,8 +23,9 @@ class IApplicationFunctions : public ServiceBase {
     kernel::Event* gpu_error_detect_event;
 
     // Commands
-    result_t PopLaunchParameter(add_service_fn_t add_service,
-                                LaunchParameterKind kind);
+    result_t PopLaunchParameter(kernel::Process* process,
+                                add_service_fn_t add_service,
+                                kernel::LaunchParameterKind kind);
     result_t EnsureSaveData(uuid_t user_id, u64* out_unknown);
     result_t GetDesiredLanguage(LanguageCode* out_language_code);
     result_t SetTerminateResult(result_t result);
