+++ src/core/horizon/services/am/common_state_getter.cpp
@@ -15,15 +15,17 @@ DEFINE_SERVICE_COMMAND_TABLE(
 result_t
 ICommonStateGetter::GetEventHandle(kernel::Process* process,
                                    OutHandle<HandleAttr::Copy> out_handle) {
-    out_handle = process->AddHandle(StateManager::GetInstance().GetMsgEvent());
+    out_handle = process->AddHandle(process->GetAppletState().GetMsgEvent());
     return RESULT_SUCCESS;
 }
 
-result_t ICommonStateGetter::ReceiveMessage(AppletMessage* out_message) {
-    const auto msg = StateManager::GetInstance().ReceiveMessage();
-    if (msg == AppletMessage::None)
+result_t
+ICommonStateGetter::ReceiveMessage(kernel::Process* process,
+                                   kernel::AppletMessage* out_message) {
+    const auto msg = process->GetAppletState().ReceiveMessage();
+    if (msg == kernel::AppletMessage::None)
         return MAKE_RESULT(Am, 0x3);
-    LOG_DEBUG(Services, "MESSAGE: {}", msg);
+    LOG_DEBUG(Services, "Message: {}", msg);
 
     *out_message = msg;
     return RESULT_SUCCESS;
@@ -49,8 +51,10 @@ result_t ICommonStateGetter::GetDefaultDisplayResolutionChangeEvent(
     return RESULT_SUCCESS;
 }
 
-result_t ICommonStateGetter::GetCurrentFocusState(AppletFocusState* out_state) {
-    *out_state = StateManager::GetInstance().GetFocusState();
+result_t
+ICommonStateGetter::GetCurrentFocusState(kernel::Process* process,
+                                         kernel::AppletFocusState* out_state) {
+    *out_state = process->GetAppletState().GetFocusState();
     return RESULT_SUCCESS;
 }
